move help text into helpers folder
